requirements.txt
PyGithub==2.1.1
anthropic>=0.39.0
python-dotenv==1.0.0
httpx>=0.27.0